Crate sqlz

This is proprietary software. If you are granted an “Evaluation License” to this library, you can develop software using this library until the specified expiry date of the license. However, to deploy or distribute the software, please purchase a “Production License” from us, which comes with technical support and maintenance for a set period.

Re-exports

pub use mssql::MsSqlPool;

pub use futures;

pub use arguments::*;

pub use column::*;

pub use connection::*;

pub use database::*;

pub use error::*;

pub use executor::*;

pub use pool::*;

pub use query::*;

pub use row::*;

pub use transaction::*;

pub use type_info::*;

Modules

arguments

column

connection

database

Traits to represent a database driver.

error

Types for working with errors produced by SQLz.

executor

mssql

pool

Provides the connection pool for asynchronous SQLz connections.

query

row

transaction

type_info