Trait sqlz::connection::ConnectOptions

pub trait ConnectOptions: 'static + Send + Sync + FromStr<Err = Error> + Debug + Clone {
    type Connection: ?Sized;

    // Required methods
    fn connect(&self) -> BoxFuture<'_, Result<Self::Connection, Error>>
       where Self::Connection: Sized;
    fn log_statements(&mut self, level: LevelFilter) -> &mut Self;
    fn log_slow_statements(
        &mut self,
        level: LevelFilter,
        duration: Duration,
    ) -> &mut Self;

    // Provided method
    fn disable_statement_logging(&mut self) -> &mut Self { ... }
}

Required Associated Types

type Connection: ?Sized

Required Methods

fn connect(&self) -> BoxFuture<'_, Result<Self::Connection, Error>>
where Self::Connection: Sized,

Establish a new database connection with the options specified by self.

fn log_statements(&mut self, level: LevelFilter) -> &mut Self

Log executed statements with the specified level

fn log_slow_statements( &mut self, level: LevelFilter, duration: Duration, ) -> &mut Self

Log executed statements with a duration above the specified duration at the specified level.

Provided Methods

fn disable_statement_logging(&mut self) -> &mut Self

Entirely disables statement logging (both slow and regular).

Object Safety

This trait is not object safe.